requirements.txt
django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
asgiref== 3.7.2
django-environ==0.9.0
PyJWT==1.7.1
pytz
tzdata==2022.6
pillow==10.2.0
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
langchain
openai
pypdf2
tiktoken
faiss-cpu
django-dotenv
python-docx
pypdf